services/requirements.txt
openai>=1.44.0
pymongo>=4.6.0
python-dotenv>=1.0.1
